+++ examples/src/main/java/filters/PXFilter.java
@@ -6,7 +6,6 @@ import com.perimeterx.models.exceptions.PXException;
 import org.springframework.stereotype.Component;
 import px.CustomBlockHandler;
 
-import javax.servlet.*;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpServletResponseWrapper;