+++ src/main/java/com/perimeterx/models/activities/PageRequestedActivityDetails.java
@@ -3,21 +3,20 @@ package com.perimeterx.models.activities;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.perimeterx.internals.cookie.RiskCookie;
 import com.perimeterx.models.PXContext;
 
 /**
  * Created by shikloshi on 07/11/2016.
  */
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public class PageRequestedActivityDetails implements ActivityDetails{
+public class PageRequestedActivityDetails implements ActivityDetails {
 
     @JsonProperty("http_method")
     private String httpMethod;
     @JsonProperty("http_version")
     private String httpVersion;
     @JsonPropertyOrder("px_cookie")
-    private RiskCookie riskCookie;
+    private String riskCookie;
 
 
     public PageRequestedActivityDetails(PXContext context) {
@@ -34,7 +33,7 @@ public class PageRequestedActivityDetails implements ActivityDetails{
         return httpVersion;
     }
 
-    public RiskCookie getRiskCookie() {
+    public String getRiskCookie() {
         return riskCookie;
     }
 }