+++ src/main/java/com/perimeterx/models/exceptions/PXCookieDecryptionException.java
@@ -1,0 +1,19 @@
+package com.perimeterx.models.exceptions;
+
+/**
+ * Created by nitzangoldfeder on 13/04/2017.
+ */
+public class PXCookieDecryptionException extends Exception {
+
+    public PXCookieDecryptionException(Throwable cause) {
+        super(cause);
+    }
+
+    public PXCookieDecryptionException(String message) {
+        super(message);
+    }
+
+    public PXCookieDecryptionException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}