+++ src/main/java/com/perimeterx/models/httpmodels/Additional.java
@@ -1,7 +1,6 @@
 package com.perimeterx.models.httpmodels;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.perimeterx.internals.cookie.RiskCookie;
 import com.perimeterx.models.PXContext;
 import com.perimeterx.models.risk.S2SCallReason;
 import com.perimeterx.utils.Constants;
@@ -13,15 +12,18 @@ import com.perimeterx.utils.Constants;
 public class Additional {
 
     @JsonProperty("px_cookie")
-    public RiskCookie PxCookie;
+    public String PxCookie;
     @JsonProperty("http_method")
     public String HttpMethod;
     @JsonProperty("http_version")
     public String HttpVersion;
     @JsonProperty("s2s_call_reason")
     public S2SCallReason CallReason;
+    @JsonProperty("px_cookie_orig")
+    public String pxCookieOrig;
     @JsonProperty("module_version")
     public final String ModuleVersion = Constants.SDK_VERSION;
+
     @JsonProperty("custom_params")
 
     public static Additional fromContext(PXContext ctx) {
@@ -30,6 +32,7 @@ public class Additional {
         additional.HttpMethod = ctx.getHttpMethod();
         additional.HttpVersion = ctx.getHttpVersion();
         additional.CallReason = ctx.getS2sCallReason();
+        additional.pxCookieOrig = ctx.getPxCookieOrig();
         return additional;
     }
 }