+++ src/test/java/com/perimeterx/internal/RiskCookieDecoderTest.java
@@ -1,7 +1,53 @@
 package com.perimeterx.internal;
 
+import com.perimeterx.api.PXConfiguration;
+import com.perimeterx.api.providers.DefaultHostnameProvider;
+import com.perimeterx.api.providers.HostnameProvider;
+import com.perimeterx.api.providers.IPProvider;
+import com.perimeterx.api.providers.RemoteAddressIPProvider;
+import com.perimeterx.internals.PXCookieValidator;
+import com.perimeterx.models.PXContext;
+import com.perimeterx.models.exceptions.PXException;
+import com.perimeterx.models.risk.S2SCallReason;
+import org.springframework.mock.web.MockHttpServletRequest;
+import org.testng.annotations.BeforeTest;
+import org.testng.annotations.Test;
+
+import javax.servlet.http.HttpServletRequest;
+
+import static org.junit.Assert.assertEquals;
+
 /**
  * Created by shikloshi on 22/11/2016.
  */
+
+@Test
 public class RiskCookieDecoderTest {
+    private final String pxCookie = "BlehldpSBV8F5ZPRa4MTbYQKbkIaXLO2AETbUSiZUobT1mvohRff7C+5EoTuKr+F1pkntRC5FGla2PWGYZiJ8g==:1000:3AWMJr6ZBQ0S6d74mSsQQJR9cWZn7LJGaSLqhX3k430YeySP71Onzxla+Qvw0PiSp9N8nIyPvYychFiG0n6ebl7eBWyWb9eTQraBzFRjWjWyuZDMr1C5e1QZ8TrMy+bUeaIuvm3J3k9pmawdHdpkyyL2HSyr35vKX5kAA1I/eWFrAETsz8UZ3xZM8cqwnq1UDQpLLWNu4ZXpDi0OLbLr+lD+mpe82Xir6TGaLVegrBZJNzJzIOhNDfpD/GXYDYpNjOtngDHWs7Se+yErPwNfmg==";
+
+    private PXConfiguration pxConfiguration;
+    private PXContext context;
+    private PXCookieValidator cookieValidator;
+
+    @BeforeTest
+    public void setUp() throws PXException {
+        HttpServletRequest request = new MockHttpServletRequest();
+        ((MockHttpServletRequest) request).addHeader("cookie", "_px=" + pxCookie);
+        IPProvider ipProvider = new RemoteAddressIPProvider();
+        HostnameProvider hostnameProvider = new DefaultHostnameProvider();
+        this.context = new PXContext(request, ipProvider, hostnameProvider, "appId");
+        this.cookieValidator = PXCookieValidator.getDecoder("cookie_token");
+        this.pxConfiguration = new PXConfiguration.Builder()
+                .cookieKey("COOKIE_KEY_STRING")
+                .appId("APP_ID")
+                .authToken("AUTH_TOKEN")
+                .build();
+    }
+
+    @Test
+    public void testOrigPXCookieSentOnCookieDecryptFails(){
+        boolean verify = cookieValidator.verify(pxConfiguration, context);
+        assertEquals(context.getS2sCallReason(), S2SCallReason.INVALID_DECRYPTION);
+        assertEquals(context.getPxCookieOrig(), pxCookie);
+    }
 }